requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
brotlipy=0.7.0=py38h2bbff1b_1003
ca-certificates=2022.3.18=haa95532_0
certifi=2021.10.8=py38haa95532_2
cffi=1.15.0=py38h2bbff1b_1
charset-normalizer=2.0.12=pypi_0
cryptography=36.0.0=py38h21b164f_0
idna=3.3=pyhd3eb1b0_0
mouseinfo=0.1.3=pypi_0
openssl=1.1.1n=h2bbff1b_0
pillow=9.0.1=pypi_0
pip=21.0.1=py38haa95532_0
pyautogui=0.9.53=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pygetwindow=0.0.9=pypi_0
pymsgbox=1.0.9=pypi_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyperclip=1.8.2=pypi_0
pyrect=0.2.0=pypi_0
pyscreeze=0.1.28=pypi_0
pysocks=1.7.1=py38haa95532_0
python=3.8.8=hdbf39b2_5
pytweening=1.0.4=pypi_0
requests=2.27.1=pypi_0
selenium=3.141.0=py38h2bbff1b_1000
setuptools=52.0.0=py38haa95532_0
sqlite=3.35.4=h2bbff1b_0
urllib3=1.26.8=pyhd3eb1b0_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
webdriver-manager=3.5.4=pypi_0
wheel=0.36.2=pyhd3eb1b0_0
win_inet_pton=1.1.0=py38haa95532_0
wincertstore=0.2=py38_0
